perf(startup): cut the static edge that re-pinned both mega-modules, and gate the budget
