+++ .github/workflows/validate.yml
@@ -572,6 +572,7 @@ jobs:
           run_step npm run validate:schemas
           run_step npm run validate:api
           run_step npm run validate:mcp
+          run_step npm run validate:startup-budget
           run_step npm run validate:mcp-route-map
           run_step npm run validate:operations
           run_step npm run validate:type-duplicates
